requirements.txt
flask
gunicorn
flask_cors
python-resize-image
tensorflow
pillow
scipy
numpy
